Refactor room page layout and fix toast close alignment
