+++ src/app/room/[roomId]/page.tsx
@@ -2,12 +2,7 @@
 
 import {
   AlertCircleIcon,
-  ArrowRightIcon,
-  CheckCircle2Icon,
-  CopyIcon,
   EyeIcon,
-  EyeOffIcon,
-  InfoIcon,
   PlayIcon,
   TargetIcon,
   TimerIcon,
@@ -17,11 +12,9 @@ import {
 import Link from "next/link";
 import { useParams, useRouter, useSearchParams } from "next/navigation";
 import {
-  type FormEvent,
   type ReactNode,
   useCallback,
   useEffect,
-  useId,
   useMemo,
   useRef,
   useState,
@@ -29,14 +22,27 @@ import {
 
 import {
   type HeaderActionState,
-  type HeaderInviteActionStatus,
   useHeaderActionRegistration,
 } from "@/components/app/HeaderActionsContext";
-import { ImageSafe } from "@/components/common/ImageSafe";
+import { FinalResultCard } from "@/components/room/FinalResultCard";
+import {
+  type HostIdentitySummary,
+  InviteDialog,
+} from "@/components/room/InviteDialog";
+import { MissingOpponentBoard } from "@/components/room/MissingOpponentBoard";
 import { ParticipantsSheet } from "@/components/room/ParticipantsSheet";
+import {
+  PlayerBoard,
+  type PlayerBoardAccent,
+} from "@/components/room/PlayerBoard";
+import {
+  RoleSelectionScreen,
+  type RoleSelectionSummary,
+} from "@/components/room/RoleSelectionScreen";
+import { RoomInformationDialog } from "@/components/room/RoomInformationDialog";
+import { formatStatusLabel } from "@/components/room/roomLabels";
 import { TargetSelectionModal } from "@/components/room/TargetSelectionModal";
 import { TurnBar } from "@/components/room/TurnBar";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import {
   Card,
@@ -45,15 +51,6 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
 import { Skeleton } from "@/components/ui/skeleton";
 import {
   canStartGame,
@@ -63,12 +60,10 @@ import {
   reduceGameState,
   selectActivePlayer,
   selectPlayers,
-  selectWinner,
 } from "@/lib/game/state";
 import {
   type Action,
   type Card as GameCard,
-  GameConclusionReason,
   type GameState,
   GameStatus,
   type Grid,
@@ -79,34 +74,25 @@ import {
   type ActionReplicator,
   createActionReplicator,
 } from "@/lib/p2p/action-sync";
-import {
-  type ConnectionPhase,
-  createPeer,
-  PeerRole,
-  type PeerRuntime,
-} from "@/lib/p2p/peer";
+import { PeerRole } from "@/lib/p2p/peer";
 import {
   GAME_PROTOCOL_NAMESPACE,
   type GameActionMessagePayload,
-  type GameProtocolMessageMap,
-  SUPPORTED_GAME_PROTOCOL_VERSIONS,
   validateGameActionMessage,
 } from "@/lib/p2p/protocol";
-import {
-  buildInviteUrl,
-  decodeGridFromToken,
-  encodeGridToToken,
-} from "@/lib/share/url";
+import { decodeGridFromToken } from "@/lib/share/url";
 import type { HostPreparationRecord } from "@/lib/storage/session";
 import { loadHostPreparation } from "@/lib/storage/session";
-import { cn, createRandomId } from "@/lib/utils";
+import { createRandomId } from "@/lib/utils";
+import {
+  type RoomPeerCreationConfig,
+  useRoomPeerRuntime,
+} from "./hooks/useRoomPeerRuntime";
 
 interface Spectator {
   id: string;
   name: string;
 }
-
-type PlayerBoardAccent = "self" | "opponent" | "neutral";
 
 type PlayerSummary = {
   player: Player;
@@ -117,1349 +103,12 @@ type PlayerSummary = {
 
 type LoadState = "idle" | "loading" | "ready" | "error";
 
-const roleLabels: Record<PlayerRole, string> = {
-  host: "Hôte",
-  guest: "Invité",
-};
-
-interface RoomPeerCreationConfig {
-  role: PeerRole;
-  peerId: string;
-  metadata: Record<string, unknown>;
-}
-
-interface RoomPeerRuntimeState {
-  runtime: PeerRuntime<GameProtocolMessageMap> | null;
-  peerId: string | null;
-  remotePeerId: string | null;
-  phase: ConnectionPhase;
-  error: Error | null;
-}
-
-const createInitialPeerState = (): Omit<RoomPeerRuntimeState, "runtime"> => ({
-  peerId: null,
-  remotePeerId: null,
-  phase: "idle",
-  error: null,
-});
-
-const useRoomPeerRuntime = (
-  config: RoomPeerCreationConfig | null,
-  remotePeerId: string | null,
-): RoomPeerRuntimeState => {
-  const runtimeRef = useRef<PeerRuntime<GameProtocolMessageMap> | null>(null);
-  const [state, setState] = useState<Omit<RoomPeerRuntimeState, "runtime">>(
-    createInitialPeerState,
-  );
-
-  useEffect(() => {
-    if (!config) {
-      if (runtimeRef.current) {
-        runtimeRef.current.destroy();
-        runtimeRef.current = null;
-      }
-      setState(createInitialPeerState());
-      return;
-    }
-
-    if (typeof window === "undefined") {
-      return;
-    }
-
-    const runtime = createPeer<GameProtocolMessageMap>({
-      role: config.role,
-      peerId: config.peerId,
-      metadata: config.metadata,
-      supportedProtocolVersions: SUPPORTED_GAME_PROTOCOL_VERSIONS,
-    });
-    runtimeRef.current = runtime;
-    setState({
-      peerId: runtime.peerId,
-      remotePeerId: runtime.remotePeerId,
-      phase: runtime.phase,
-      error: null,
-    });
-
-    const offPeerOpen = runtime.events.on("peer/open", ({ peerId }) => {
-      setState((previous) => ({ ...previous, peerId }));
-    });
-    const offRemote = runtime.events.on("connection/remote", ({ peerId }) => {
-      setState((previous) => ({ ...previous, remotePeerId: peerId }));
-    });
-    const offPhase = runtime.events.on("connection/phase", ({ phase }) => {
-      setState((previous) => ({ ...previous, phase }));
-    });
-    const offError = runtime.events.on("connection/error", ({ error }) => {
-      setState((previous) => ({ ...previous, error, phase: "error" }));
-    });
-    const offDisconnected = runtime.events.on("connection/disconnected", () => {
-      setState((previous) => ({
-        ...previous,
-        remotePeerId: null,
-        phase: "reconnecting",
-      }));
-    });
-    const offPeerClose = runtime.events.on("peer/close", () => {
-      setState(createInitialPeerState());
-    });
-
-    return () => {
-      offPeerOpen();
-      offRemote();
-      offPhase();
-      offError();
-      offDisconnected();
-      offPeerClose();
-      runtime.destroy();
-      runtimeRef.current = null;
-      setState(createInitialPeerState());
-    };
-  }, [config]);
-
-  useEffect(() => {
-    if (!config || config.role !== PeerRole.Guest) {
-      return;
-    }
-    const runtime = runtimeRef.current;
-    if (!runtime) {
-      return;
-    }
-    if (!remotePeerId) {
-      runtime.disconnect();
-      return;
-    }
-    runtime.connect(remotePeerId).catch((error) => {
-      setState((previous) => ({ ...previous, error, phase: "error" }));
-    });
-  }, [config, remotePeerId]);
-
-  return {
-    runtime: runtimeRef.current,
-    peerId: state.peerId,
-    remotePeerId: state.remotePeerId,
-    phase: state.phase,
-    error: state.error,
-  };
-};
-
 interface InviteContext {
   hostId: string;
   hostName: string;
   grid: Grid;
 }
 
-const accentClasses: Record<PlayerBoardAccent, string> = {
-  self: "border-sky-400/80 bg-sky-500/10 dark:border-sky-500/60 dark:bg-sky-500/15",
-  opponent:
-    "border-rose-400/80 bg-rose-500/10 dark:border-rose-500/60 dark:bg-rose-500/15",
-  neutral: "border-border/70 bg-muted/40",
-};
-
-const formatStatusLabel = (status: GameStatus): string => {
-  switch (status) {
-    case GameStatus.Idle:
-      return "Initialisation";
-    case GameStatus.Lobby:
-      return "Préparation de la partie";
-    case GameStatus.Playing:
-      return "En cours";
-    case GameStatus.Finished:
-      return "Partie terminée";
-    default:
-      return status;
-  }
-};
-
-const getConclusionLabel = (reason: GameConclusionReason): string => {
-  switch (reason) {
-    case GameConclusionReason.CorrectGuess:
-      return "Victoire sur bonne réponse";
-    case GameConclusionReason.IncorrectGuess:
-      return "Victoire par mauvaise réponse";
-    default:
-      return reason;
-  }
-};
-
-type CopyStatus = HeaderInviteActionStatus;
-
-interface HostIdentitySummary {
-  id: string;
-  name: string;
-}
-
-function InviteDialog({
-  grid,
-  roomId,
-  host,
-  canShare,
-  allowJoin,
-  canJoinAsPlayer,
-  onJoin,
-  isJoining,
-}: {
-  grid: Grid | null;
-  roomId: string | null;
-  host: HostIdentitySummary | null;
-  canShare: boolean;
-  allowJoin: boolean;
-  canJoinAsPlayer: boolean;
-  onJoin: (nickname: string) => void;
-  isJoining: boolean;
-}) {
-  const [origin, setOrigin] = useState<string | null>(null);
-  const [copyStatus, setCopyStatus] = useState<CopyStatus>("idle");
-  const [copyError, setCopyError] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (typeof window === "undefined") {
-      return;
-    }
-    setOrigin(window.location.origin);
-  }, []);
-
-  const sharePayload = useMemo(() => {
-    if (!grid) {
-      return { token: null as string | null, error: null as string | null };
-    }
-    try {
-      const token = encodeGridToToken(grid);
-      return { token, error: null };
-    } catch (error) {
-      console.error("Impossible de générer le code d’invitation.", error);
-      return {
-        token: null,
-        error:
-          error instanceof Error
-            ? error.message
-            : "Erreur inconnue lors de la génération du code.",
-      };
-    }
-  }, [grid]);
-
-  const inviteUrl = useMemo(() => {
-    if (!sharePayload.token || !origin || !roomId || !host) {
-      return null;
-    }
-    try {
-      return buildInviteUrl(
-        origin,
-        { roomId, hostId: host.id, hostName: host.name },
-        sharePayload.token,
-      );
-    } catch (error) {
-      console.error("Impossible de construire l’URL d’invitation.", error);
-      return null;
-    }
-  }, [origin, sharePayload.token, roomId, host]);
-
-  useEffect(() => {
-    if (inviteUrl === null) {
-      setCopyStatus("idle");
-      setCopyError(null);
-      return;
-    }
-    setCopyStatus("idle");
-    setCopyError(null);
-  }, [inviteUrl]);
-
-  useEffect(() => {
-    if (copyStatus !== "copied") {
-      return;
-    }
-    const timeout = window.setTimeout(() => {
-      setCopyStatus("idle");
-    }, 2500);
-    return () => window.clearTimeout(timeout);
-  }, [copyStatus]);
-
-  const handleCopy = useCallback(async () => {
-    if (!inviteUrl) {
-      return;
-    }
-    setCopyError(null);
-    setCopyStatus("pending");
-    if (
-      typeof navigator === "undefined" ||
-      !navigator.clipboard ||
-      typeof navigator.clipboard.writeText !== "function"
-    ) {
-      setCopyStatus("error");
-      setCopyError(
-        "Copie automatique indisponible : sélectionnez le lien et copiez-le manuellement.",
-      );
-      return;
-    }
-    try {
-      await navigator.clipboard.writeText(inviteUrl);
-      setCopyStatus("copied");
-      setCopyError(null);
-    } catch (error) {
-      console.error("Impossible de copier le lien d’invitation.", error);
-      setCopyStatus("error");
-      setCopyError(
-        "Impossible de copier automatiquement le lien. Sélectionnez-le et copiez-le manuellement.",
-      );
-    }
-  }, [inviteUrl]);
-
-  const inviteHeaderAction = useMemo<HeaderActionState>(() => {
-    if (!canShare) {
-      return {};
-    }
-    const hasInviteGenerationError = Boolean(sharePayload.error);
-    const isInviteAvailable = Boolean(inviteUrl);
-    if (!hasInviteGenerationError && !isInviteAvailable) {
-      return {};
-    }
-
-    return {
-      invite: {
-        label: hasInviteGenerationError
-          ? "Lien indisponible"
-          : "Copier le lien",
-        ariaLabel: hasInviteGenerationError
-          ? "Le lien d’invitation est actuellement indisponible"
-          : "Copier le lien d’invitation de cette salle",
-        icon: CopyIcon,
-        onActivate: handleCopy,
-        status: hasInviteGenerationError ? "error" : copyStatus,
-        disabled:
-          hasInviteGenerationError ||
-          !isInviteAvailable ||
-          copyStatus === "pending",
-        pendingLabel: "Copie en cours…",
-        successLabel: "Lien copié",
-        errorLabel: hasInviteGenerationError ? "Lien indisponible" : undefined,
-        feedbackMessage:
-          copyStatus === "copied"
-            ? "Lien d’invitation copié dans le presse-papier."
-            : copyStatus === "error"
-              ? (copyError ??
-                "Impossible de copier automatiquement le lien d’invitation.")
-              : hasInviteGenerationError
-                ? `Impossible de générer le lien d’invitation : ${sharePayload.error}`
-                : null,
-        feedbackTone:
-          copyStatus === "error" || hasInviteGenerationError
-            ? "assertive"
-            : "polite",
-      },
-    };
-  }, [
-    canShare,
-    copyError,
-    copyStatus,
-    handleCopy,
-    inviteUrl,
-    sharePayload.error,
-  ]);
-  useHeaderActionRegistration(inviteHeaderAction);
-
-  if (!grid) {
-    return null;
-  }
-
-  const hasTokenError = Boolean(sharePayload.error);
-  const invitePlaceholder = hasTokenError
-    ? "Erreur lors de la génération du lien"
-    : "Le lien apparaîtra ici dès qu’il est prêt.";
-  const hostName = host?.name ?? "Hôte";
-  const isSubmitting = isJoining;
-
-  return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <Button
-          type="button"
-          variant="outline"
-          size="sm"
-          className="flex items-center gap-2"
-        >
-          <CopyIcon aria-hidden className="size-4" />
-          Inviter
-        </Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-xl">
-        <DialogHeader>
-          <DialogTitle>Inviter un adversaire</DialogTitle>
-          <DialogDescription>
-            Partagez ce lien pour permettre à un joueur de rejoindre la salle et
-            importer automatiquement le plateau « {grid.name} ».
-          </DialogDescription>
-        </DialogHeader>
-        <div className="space-y-6">
-          <div className="space-y-2">
-            <label
-              htmlFor="invite-link"
-              className="text-sm font-medium text-foreground"
-            >
-              Lien d’invitation
-            </label>
-            <div className="flex flex-col gap-2 sm:flex-row sm:items-stretch">
-              <input
-                id="invite-link"
-                type="text"
-                value={inviteUrl ?? ""}
-                readOnly
-                placeholder={invitePlaceholder}
-                onFocus={(event) => event.currentTarget.select()}
-                spellCheck={false}
-                className="w-full flex-1 rounded-md border border-border/70 bg-background px-3 py-2 text-xs font-mono shadow-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring sm:text-sm"
-                aria-invalid={hasTokenError || copyStatus === "error"}
-                aria-describedby={
-                  hasTokenError
-                    ? "invite-link-error"
-                    : copyStatus === "error" && copyError
-                      ? "invite-link-copy-error"
-                      : undefined
-                }
-                disabled={!canShare}
-              />
-              <Button
-                type="button"
-                variant="outline"
-                onClick={handleCopy}
-                disabled={!inviteUrl || !canShare || copyStatus === "pending"}
-                className="sm:w-auto sm:flex-none"
-              >
-                {copyStatus === "copied" ? (
-                  <>
-                    <CheckCircle2Icon aria-hidden className="mr-2 size-4" />
-                    Lien copié
-                  </>
-                ) : copyStatus === "pending" ? (
-                  <>
-                    <TimerIcon
-                      aria-hidden
-                      className="mr-2 size-4 animate-spin"
-                    />
-                    Copie…
-                  </>
-                ) : (
-                  <>
-                    <CopyIcon aria-hidden className="mr-2 size-4" />
-                    Copier
-                  </>
-                )}
-              </Button>
-            </div>
-            {sharePayload.token ? (
-              <p className="text-xs text-muted-foreground">
-                Code compressé :{" "}
-                <code className="break-all rounded bg-muted px-1 py-0.5">
-                  {sharePayload.token}
-                </code>
-              </p>
-            ) : null}
-            {hasTokenError ? (
-              <div
-                id="invite-link-error"
-                className="flex items-start gap-2 rounded-md border border-destructive/40 bg-destructive/10 px-3 py-2 text-sm text-destructive"
-                aria-live="polite"
-              >
-                <AlertCircleIcon aria-hidden className="mt-0.5 size-4" />
-                <span>
-                  Impossible de générer le lien d’invitation.{" "}
-                  {sharePayload.error}
-                </span>
-              </div>
-            ) : canShare ? (
-              <p className="flex items-center gap-2 text-sm text-muted-foreground">
-                <ArrowRightIcon aria-hidden className="size-4" />
-                Les invités peuvent ouvrir ce lien sur n’importe quel appareil
-                pour importer automatiquement votre plateau.
-              </p>
-            ) : (
-              <p className="flex items-center gap-2 text-sm text-muted-foreground">
-                <EyeIcon aria-hidden className="size-4" />
-                Seul l’hôte peut partager ce lien depuis cet appareil.
-              </p>
-            )}
-            {copyStatus === "error" && copyError ? (
-              <p
-                id="invite-link-copy-error"
-                className="text-sm text-destructive"
-                aria-live="polite"
-              >
-                {copyError}
-              </p>
-            ) : null}
-          </div>
-          {allowJoin ? (
-            <div className="space-y-4 border-t border-border/60 pt-4">
-              <div className="space-y-1">
-                <h3 className="text-sm font-semibold text-foreground">
-                  Rejoindre la salle de {hostName}
-                </h3>
-                <p className="text-sm text-muted-foreground">
-                  Sélectionnez votre rôle puis choisissez un pseudo si vous
-                  souhaitez affronter l’hôte.
-                </p>
-                {!canJoinAsPlayer ? (
-                  <p className="text-xs text-muted-foreground">
-                    La place de joueur est déjà occupée. Vous pouvez néanmoins
-                    observer la partie en tant que spectateur.
-                  </p>
-                ) : null}
-              </div>
-              <JoinAsGuestForm
-                hostName={hostName}
-                onJoin={onJoin}
-                disabled={!allowJoin || isJoining}
-                isSubmitting={isSubmitting}
-                canJoinAsPlayer={canJoinAsPlayer}
-              />
-            </div>
-          ) : null}
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-type JoinRole = "player" | "spectator";
-
-function JoinAsGuestForm({
-  hostName,
-  onJoin,
-  disabled,
-  isSubmitting,
-  canJoinAsPlayer,
-}: {
-  hostName: string;
-  onJoin: (nickname: string) => void;
-  disabled: boolean;
-  isSubmitting: boolean;
-  canJoinAsPlayer: boolean;
-}) {
-  const [nickname, setNickname] = useState("");
-  const [localError, setLocalError] = useState<string | null>(null);
-  const [selectedRole, setSelectedRole] = useState<JoinRole>(
-    canJoinAsPlayer ? "player" : "spectator",
-  );
-  const roleFieldId = useId();
-  const playerOptionId = `${roleFieldId}-player`;
-  const spectatorOptionId = `${roleFieldId}-spectator`;
-
-  useEffect(() => {
-    if (!canJoinAsPlayer) {
-      setSelectedRole("spectator");
-    }
-  }, [canJoinAsPlayer]);
-
-  const handleRoleChange = useCallback((role: JoinRole) => {
-    setSelectedRole(role);
-    setLocalError(null);
-  }, []);
-
-  const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (disabled || selectedRole !== "player") {
-      return;
-    }
-    const trimmed = nickname.trim();
-    if (!trimmed) {
-      setLocalError("Veuillez renseigner un pseudo pour rejoindre la salle.");
-      return;
-    }
-    try {
-      onJoin(trimmed);
-      setLocalError(null);
-    } catch (error) {
-      setLocalError(
-        error instanceof Error
-          ? error.message
-          : "Impossible de rejoindre la salle. Réessayez dans un instant.",
-      );
-    }
-  };
-
-  return (
-    <form className="space-y-4" onSubmit={handleSubmit}>
-      <fieldset className="space-y-3">
-        <legend className="text-sm font-medium text-foreground">
-          Choisissez comment participer
-        </legend>
-        <div className="space-y-2">
-          <label
-            htmlFor={playerOptionId}
-            className={cn(
-              "flex items-start gap-3 rounded-md border px-3 py-3 text-sm transition-colors",
-              selectedRole === "player"
-                ? "border-primary bg-primary/5 ring-2 ring-primary/30"
-                : "border-border/70 bg-background",
-              canJoinAsPlayer && !disabled
-                ? "cursor-pointer hover:border-primary/50 focus-within:border-primary focus-within:ring-2 focus-within:ring-primary/20"
-                : "cursor-not-allowed opacity-60",
-            )}
-          >
-            <input
-              id={playerOptionId}
-              type="radio"
-              name={`${roleFieldId}-role`}
-              value="player"
-              checked={selectedRole === "player"}
-              onChange={() => handleRoleChange("player")}
-              disabled={disabled || isSubmitting || !canJoinAsPlayer}
-              className="mt-1 h-4 w-4 border-border/70 text-primary focus:ring-ring"
-            />
-            <div className="flex flex-col gap-1">
-              <span className="font-medium text-foreground">Joueur</span>
-              <span className="text-xs text-muted-foreground">
-                Défiez l’hôte sur ce plateau personnalisé.
-              </span>
-              {!canJoinAsPlayer ? (
-                <span className="text-xs text-destructive">
-                  La place de joueur est déjà occupée.
-                </span>
-              ) : null}
-            </div>
-          </label>
-          <label
-            htmlFor={spectatorOptionId}
-            className={cn(
-              "flex items-start gap-3 rounded-md border px-3 py-3 text-sm transition-colors",
-              selectedRole === "spectator"
-                ? "border-primary bg-primary/5 ring-2 ring-primary/30"
-                : "border-border/70 bg-background",
-              "cursor-pointer hover:border-primary/50 focus-within:border-primary focus-within:ring-2 focus-within:ring-primary/20",
-            )}
-          >
-            <input
-              id={spectatorOptionId}
-              type="radio"
-              name={`${roleFieldId}-role`}
-              value="spectator"
-              checked={selectedRole === "spectator"}
-              onChange={() => handleRoleChange("spectator")}
-              className="mt-1 h-4 w-4 border-border/70 text-primary focus:ring-ring"
-            />
-            <div className="flex flex-col gap-1">
-              <span className="font-medium text-foreground">Spectateur</span>
-              <span className="text-xs text-muted-foreground">
-                Observez la partie en direct sans interagir avec le plateau.
-              </span>
-            </div>
-          </label>
-        </div>
-      </fieldset>
-
-      {selectedRole === "player" ? (
-        <>
-          <div className="space-y-2">
-            <label
-              htmlFor="guest-nickname"
-              className="text-sm font-medium text-foreground"
-            >
-              Votre pseudo
-            </label>
-            <input
-              id="guest-nickname"
-              type="text"
-              value={nickname}
-              onChange={(event) => {
-                setNickname(event.target.value);
-                if (localError) {
-                  setLocalError(null);
-                }
-              }}
-              onBlur={() => {
-                if (!nickname.trim()) {
-                  setLocalError(
-                    "Un pseudo est nécessaire pour identifier chaque joueur.",
-                  );
-                }
-              }}
-              className="w-full rounded-md border border-border/70 bg-background px-3 py-2 text-sm shadow-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-              placeholder="Invité mystère"
-              maxLength={40}
-              autoComplete="off"
-              disabled={disabled || isSubmitting}
-            />
-          </div>
-          <div className="space-y-2 text-sm text-muted-foreground">
-            <p>
-              Ce pseudo sera visible par les autres participants afin que votre
-              adversaire sache qui a rejoint la salle.
-            </p>
-            <p>
-              Vous pourrez le modifier ultérieurement depuis votre tableau de
-              bord.
-            </p>
-          </div>
-        </>
-      ) : (
-        <div className="space-y-2 text-sm text-muted-foreground">
-          <p>Aucun pseudo n’est requis pour observer la partie.</p>
-          <p>
-            Fermez cette fenêtre pour suivre la partie en tant que spectateur.
-          </p>
-        </div>
-      )}
-
-      {localError ? (
-        <div className="rounded-md border border-destructive/40 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-          {localError}
-        </div>
-      ) : null}
-
-      {selectedRole === "player" ? (
-        <Button
-          type="submit"
-          disabled={disabled || isSubmitting || !canJoinAsPlayer}
-        >
-          <UsersIcon aria-hidden className="mr-2 size-4" />
-          {isSubmitting ? "Connexion…" : `Rejoindre ${hostName}`}
-        </Button>
-      ) : (
-        <DialogClose asChild>
-          <Button type="button" variant="outline">
-            <EyeIcon aria-hidden className="mr-2 size-4" />
-            Observer la partie
-          </Button>
-        </DialogClose>
-      )}
-    </form>
-  );
-}
-function RoomInformationDialog({
-  roomId,
-  grid,
-  status,
-  hostName,
-}: {
-  roomId: string;
-  grid: Grid;
-  status: GameStatus;
-  hostName: string | null;
-}) {
-  return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <Button
-          type="button"
-          variant="outline"
-          size="sm"
-          className="flex items-center gap-2"
-        >
-          <InfoIcon aria-hidden className="size-4" />
-          Infos salle
-        </Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-lg">
-        <DialogHeader>
-          <DialogTitle>Informations sur la salle</DialogTitle>
-          <DialogDescription>
-            Récapitulatif du plateau et des paramètres de cette partie.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="space-y-5 text-sm">
-          <div className="grid gap-3">
-            <div className="space-y-1">
-              <span className="text-xs font-semibold uppercase text-muted-foreground">
-                Identifiant
-              </span>
-              <code className="inline-flex items-center rounded bg-muted px-2 py-1 text-xs font-mono">
-                {roomId}
-              </code>
-            </div>
-            <div className="space-y-1">
-              <span className="text-xs font-semibold uppercase text-muted-foreground">
-                Plateau
-              </span>
-              <p className="text-foreground">
-                {grid.name} — {grid.rows} × {grid.columns} ({grid.cards.length}{" "}
-                cartes)
-              </p>
-            </div>
-            <div className="space-y-1">
-              <span className="text-xs font-semibold uppercase text-muted-foreground">
-                Hôte
-              </span>
-              <p className="text-foreground">{hostName ?? "Hôte inconnu"}</p>
-            </div>
-            <div className="space-y-1">
-              <span className="text-xs font-semibold uppercase text-muted-foreground">
-                État
-              </span>
-              <p className="text-foreground">{formatStatusLabel(status)}</p>
-            </div>
-          </div>
-          <p className="text-muted-foreground">
-            La salle reste active tant que les participants conservent cette
-            page ouverte. Les échanges se synchronisent directement entre
-            appareils via une connexion chiffrée de navigateur à navigateur.
-          </p>
-        </div>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function SecretCardPreview({ card }: { card: GameCard }) {
-  return (
-    <div className="flex items-start gap-3 rounded-md border border-border/60 bg-background/80 p-3 shadow-sm">
-      <div className="relative h-16 w-16 overflow-hidden rounded-md bg-muted">
-        {card.imageUrl ? (
-          <ImageSafe
-            src={card.imageUrl}
-            alt={card.label}
-            className="absolute inset-0"
-            imageProps={{ sizes: "64px" }}
-            fallback={
-              <span className="absolute inset-0 flex items-center justify-center px-1 text-center text-xs text-destructive">
-                Image indisponible
-              </span>
-            }
-          />
-        ) : (
-          <span className="absolute inset-0 flex items-center justify-center px-2 text-center text-xs text-muted-foreground">
-            Aucun visuel
-          </span>
-        )}
-      </div>
-      <div className="flex-1">
-        <p className="text-sm font-medium text-foreground">{card.label}</p>
-        {card.description ? (
-          <p className="text-xs text-muted-foreground">{card.description}</p>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
-interface CardTileProps {
-  card: GameCard;
-  hidden: boolean;
-  showSecretBadge: boolean;
-  disabled: boolean;
-  onClick?: () => void;
-  actionLabel?: string;
-  interactionType: "toggle" | "none";
-}
-
-function CardTile({
-  card,
-  hidden,
-  showSecretBadge,
-  disabled,
-  onClick,
-  actionLabel,
-  interactionType,
-}: CardTileProps) {
-  const label = actionLabel ?? card.label;
-  const ariaPressed = interactionType === "toggle" ? hidden : undefined;
-
-  const content = (
-    <div className="flex flex-col gap-2 rounded-md border border-border/60 bg-background/90 p-3 shadow-sm transition-colors hover:border-border">
-      <div className="relative aspect-square overflow-hidden rounded-md bg-muted">
-        {card.imageUrl ? (
-          <ImageSafe
-            src={card.imageUrl}
-            alt={card.label}
-            className="absolute inset-0"
-            imageProps={{
-              sizes: "(min-width: 1280px) 180px, (min-width: 768px) 33vw, 50vw",
-            }}
-            fallback={
-              <span className="absolute inset-0 flex items-center justify-center px-2 text-center text-xs text-destructive">
-                Image indisponible
-              </span>
-            }
-          />
-        ) : (
-          <span className="absolute inset-0 flex items-center justify-center px-2 text-center text-xs text-muted-foreground">
-            Aucun visuel
-          </span>
-        )}
-        <div className="absolute left-2 top-2 flex items-center gap-1 rounded-full bg-background/95 px-2 py-1 text-xs font-medium text-foreground shadow-sm">
-          {hidden ? (
-            <EyeOffIcon aria-hidden className="size-3.5" />
-          ) : (
-            <EyeIcon aria-hidden className="size-3.5" />
-          )}
-          <span>{hidden ? "Masquée" : "Visible"}</span>
-        </div>
-        {showSecretBadge ? (
-          <div className="absolute bottom-2 right-2 flex items-center gap-1 rounded-full bg-primary px-2 py-1 text-xs font-semibold text-primary-foreground shadow">
-            <TargetIcon aria-hidden className="size-3.5" />
-            Carte secrète
-          </div>
-        ) : null}
-        {hidden ? (
-          <div
-            className="absolute inset-0 bg-background/60 backdrop-blur-[1px]"
-            aria-hidden
-          />
-        ) : null}
-      </div>
-      <div className="space-y-1">
-        <p className="text-sm font-medium text-foreground">{card.label}</p>
-        {card.description ? (
-          <p className="text-xs text-muted-foreground">{card.description}</p>
-        ) : null}
-      </div>
-    </div>
-  );
-
-  return (
-    <button
-      type="button"
-      className={cn(
-        "w-full text-left",
-        disabled
-          ? "cursor-default"
-          : "cursor-pointer focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
-      )}
-      disabled={disabled}
-      onClick={disabled ? undefined : onClick}
-      aria-label={label}
-      aria-pressed={ariaPressed}
-    >
-      {content}
-    </button>
-  );
-}
-interface PlayerBoardProps {
-  player: Player;
-  grid: Grid;
-  accent: PlayerBoardAccent;
-  hiddenCardIds: Set<string>;
-  secretCard: GameCard | null;
-  showSecretCard: boolean;
-  allowSecretSelection: boolean;
-  allowCardToggle: boolean;
-  onRequestSecretSelection?: () => void;
-  onToggleCard(cardId: string): void;
-  status: GameStatus;
-  isLocal: boolean;
-  isActiveTurn: boolean;
-  isSpectatorView: boolean;
-  ready: boolean;
-}
-
-function PlayerBoard({
-  player,
-  grid,
-  accent,
-  hiddenCardIds,
-  secretCard,
-  showSecretCard,
-  allowSecretSelection,
-  allowCardToggle,
-  onRequestSecretSelection,
-  onToggleCard,
-  status,
-  isLocal,
-  isActiveTurn,
-  isSpectatorView,
-  ready,
-}: PlayerBoardProps) {
-  const templateColumns = useMemo(
-    () => `repeat(${grid.columns}, minmax(0, 1fr))`,
-    [grid.columns],
-  );
-
-  const interactionType: "toggle" | "none" = allowCardToggle
-    ? "toggle"
-    : "none";
-
-  const hiddenCount = hiddenCardIds.size;
-
-  let secretContent: React.ReactNode;
-  if (secretCard && showSecretCard) {
-    secretContent = <SecretCardPreview card={secretCard} />;
-  } else if (secretCard) {
-    secretContent = (
-      <p className="text-sm text-muted-foreground">
-        {isSpectatorView
-          ? `Carte secrète sélectionnée : ${secretCard.label}.`
-          : "Carte secrète sélectionnée et cachée pour l’adversaire."}
-      </p>
-    );
-  } else if (allowSecretSelection) {
-    secretContent = (
-      <p className="text-sm text-muted-foreground">
-        Choisissez une carte secrète à l’aide du sélecteur dédié pour pouvoir
-        démarrer la partie.
-      </p>
-    );
-  } else {
-    secretContent = (
-      <p className="text-sm text-muted-foreground">
-        En attente que {player.name} choisisse sa carte secrète.
-      </p>
-    );
-  }
-
-  return (
-    <Card
-      className={cn(
-        "flex h-full min-h-0 flex-col border-2 shadow-sm transition-colors",
-        accentClasses[accent],
-      )}
-    >
-      <CardHeader className="space-y-3">
-        <div className="flex flex-wrap items-center justify-between gap-3">
-          <CardTitle className="text-lg font-semibold text-foreground">
-            {isLocal ? "Votre plateau" : `Plateau de ${player.name}`}
-          </CardTitle>
-          <div className="flex flex-wrap items-center gap-2">
-            <Badge
-              variant="outline"
-              className="border-border/60 text-xs font-semibold uppercase tracking-wide text-muted-foreground"
-            >
-              {roleLabels[player.role]}
-            </Badge>
-            {isLocal ? (
-              <Badge className="bg-sky-500/15 text-sky-600 dark:text-sky-300">
-                Vous
-              </Badge>
-            ) : null}
-            {ready ? (
-              <Badge className="bg-emerald-500/20 text-emerald-600 dark:text-emerald-300">
-                Prêt
-              </Badge>
-            ) : (
-              <Badge
-                variant="outline"
-                className="border-dashed text-muted-foreground"
-              >
-                À préparer
-              </Badge>
-            )}
-            {isActiveTurn ? (
-              <Badge className="flex items-center gap-1 bg-primary/15 text-primary">
-                <TimerIcon aria-hidden className="size-3.5" />
-                Tour en cours
-              </Badge>
-            ) : null}
-          </div>
-        </div>
-        <CardDescription>
-          {grid.rows} × {grid.columns} — {grid.cards.length} cartes
-        </CardDescription>
-      </CardHeader>
-      <CardContent className="flex flex-1 min-h-0 flex-col space-y-6 overflow-hidden">
-        <div className="space-y-3">
-          <div className="flex flex-wrap items-center gap-3">
-            <h3 className="flex items-center gap-2 text-sm font-semibold text-foreground">
-              <TargetIcon aria-hidden className="size-4 text-primary" />
-              Carte secrète
-            </h3>
-            {allowSecretSelection && onRequestSecretSelection ? (
-              <Button
-                type="button"
-                size="sm"
-                onClick={onRequestSecretSelection}
-              >
-                {secretCard ? "Modifier" : "Choisir"} la carte secrète
-              </Button>
-            ) : null}
-          </div>
-          {secretContent}
-        </div>
-
-        <div className="flex flex-1 min-h-0 flex-col space-y-3">
-          <div className="flex flex-wrap items-center justify-between text-xs text-muted-foreground">
-            <span>
-              {hiddenCount} carte{hiddenCount > 1 ? "s" : ""} masquée
-              {hiddenCount > 1 ? "s" : ""}
-            </span>
-            {status === GameStatus.Playing ? (
-              <span>
-                {isActiveTurn
-                  ? "Vous pouvez manipuler vos cartes pendant ce tour."
-                  : isLocal
-                    ? "Tour de l’adversaire."
-                    : "Tour en cours."}
-              </span>
-            ) : null}
-          </div>
-          <div className="min-h-0 flex-1 overflow-hidden">
-            <ul
-              className="grid h-full list-none gap-2 overflow-auto pr-1 sm:gap-3"
-              style={{ gridTemplateColumns: templateColumns }}
-              aria-label={`Plateau de ${player.name}`}
-            >
-              {grid.cards.map((card) => {
-                const hidden = hiddenCardIds.has(card.id);
-                const showSecret =
-                  showSecretCard && player.secretCardId === card.id;
-                const disabled = interactionType === "none";
-                const actionLabel =
-                  interactionType === "toggle"
-                    ? hidden
-                      ? `Révéler ${card.label}`
-                      : `Masquer ${card.label}`
-                    : card.label;
-                const handleClick =
-                  interactionType === "toggle"
-                    ? () => onToggleCard(card.id)
-                    : undefined;
-                return (
-                  <li key={card.id} className="list-none">
-                    <CardTile
-                      card={card}
-                      hidden={hidden}
-                      showSecretBadge={showSecret}
-                      disabled={disabled}
-                      onClick={handleClick}
-                      actionLabel={actionLabel}
-                      interactionType={interactionType}
-                    />
-                  </li>
-                );
-              })}
-            </ul>
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-}
-
-interface RoleSelectionSummary {
-  name: string;
-  details: string;
-}
-
-interface RoleSelectionScreenProps {
-  gridSummary: RoleSelectionSummary | null;
-  isAlreadyPlayer: boolean;
-  isViewingAsSpectator: boolean;
-  playerName: string | null;
-  playerRole: PlayerRole | null;
-  nickname: string;
-  canJoinAsPlayer: boolean;
-  isJoining: boolean;
-  error: string | null;
-  onNicknameChange(value: string): void;
-  onConfirmPlayer(): void;
-  onConfirmSpectator(): void;
-}
-
-function RoleSelectionScreen({
-  gridSummary,
-  isAlreadyPlayer,
-  isViewingAsSpectator,
-  playerName,
-  playerRole,
-  nickname,
-  canJoinAsPlayer,
-  isJoining,
-  error,
-  onNicknameChange,
-  onConfirmPlayer,
-  onConfirmSpectator,
-}: RoleSelectionScreenProps) {
-  const trimmedNickname = nickname.trim();
-  const playerLabel = playerRole ? roleLabels[playerRole] : null;
-  const joinDisabled =
-    !canJoinAsPlayer || trimmedNickname.length === 0 || isJoining;
-
-  const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    if (joinDisabled) {
-      return;
-    }
-    onConfirmPlayer();
-  };
-
-  return (
-    <div className="full-height-page flex min-h-0 flex-1 flex-col items-center justify-center gap-10 py-12">
-      <section className="flex flex-col items-center gap-3 text-center">
-        <div className="inline-flex items-center gap-2 rounded-full bg-primary/10 px-3 py-1 text-sm font-medium text-primary">
-          <UsersIcon aria-hidden className="size-4" />
-          Sélection du rôle
-        </div>
-        <h1 className="text-3xl font-semibold tracking-tight sm:text-4xl">
-          Comment souhaitez-vous participer ?
-        </h1>
-        <p className="max-w-2xl text-base text-muted-foreground sm:text-lg">
-          Choisissez de rejoindre la salle comme joueur actif ou d’observer la
-          partie sans manipuler les plateaux.
-        </p>
-        {gridSummary ? (
-          <p className="text-sm text-muted-foreground">
-            Plateau « {gridSummary.name} » — {gridSummary.details}
-          </p>
-        ) : null}
-      </section>
-
-      <div className="grid w-full max-w-4xl gap-6 md:grid-cols-2">
-        <Card className="border border-border/70">
-          <CardHeader>
-            <CardTitle>Rejoindre en tant que joueur</CardTitle>
-            <CardDescription>
-              Affrontez votre adversaire en manipulant votre plateau dédié.
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            {isAlreadyPlayer ? (
-              <>
-                <p className="text-sm text-muted-foreground">
-                  Vous participez déjà en tant que
-                  {playerLabel ? ` ${playerLabel.toLowerCase()}` : " joueur"}
-                  {playerName ? ` « ${playerName} »` : ""}.
-                </p>
-                <p className="text-sm text-muted-foreground">
-                  {isViewingAsSpectator
-                    ? "Sélectionnez « Revenir en tant que joueur » pour reprendre le contrôle de votre plateau sur cet appareil."
-                    : "Sélectionnez « Continuer en tant que joueur » pour accéder à votre plateau et finaliser la préparation."}
-                </p>
-                <Button type="button" onClick={onConfirmPlayer}>
-                  <UsersIcon aria-hidden className="mr-2 size-4" />
-                  {isViewingAsSpectator
-                    ? "Revenir en tant que joueur"
-                    : "Continuer en tant que joueur"}
-                </Button>
-              </>
-            ) : (
-              <>
-                <p className="text-sm text-muted-foreground">
-                  Prenez la prochaine place disponible pour affronter l’hôte.
-                </p>
-                <form className="space-y-4" onSubmit={handleSubmit}>
-                  <div className="space-y-2">
-                    <label
-                      htmlFor="role-selection-nickname"
-                      className="text-sm font-medium text-foreground"
-                    >
-                      Votre pseudo
-                    </label>
-                    <input
-                      id="role-selection-nickname"
-                      type="text"
-                      value={nickname}
-                      onChange={(event) => onNicknameChange(event.target.value)}
-                      className="w-full rounded-md border border-border/70 bg-background px-3 py-2 text-sm shadow-sm focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-                      placeholder="Invité mystère"
-                      maxLength={40}
-                      autoComplete="off"
-                      disabled={!canJoinAsPlayer || isJoining}
-                    />
-                    <p className="text-xs text-muted-foreground">
-                      Ce pseudo identifie votre plateau pendant la partie.
-                    </p>
-                  </div>
-                  {error ? (
-                    <div className="rounded-md border border-destructive/40 bg-destructive/10 px-3 py-2 text-sm text-destructive">
-                      {error}
-                    </div>
-                  ) : null}
-                  {!canJoinAsPlayer ? (
-                    <div className="rounded-md border border-border/60 bg-muted/30 px-3 py-2 text-sm text-muted-foreground">
-                      Les deux places de joueur sont déjà occupées. Choisissez
-                      le mode spectateur pour suivre la partie.
-                    </div>
-                  ) : null}
-                  <Button type="submit" disabled={joinDisabled}>
-                    <UsersIcon aria-hidden className="mr-2 size-4" />
-                    {isJoining ? "Connexion…" : "Rejoindre la partie"}
-                  </Button>
-                </form>
-              </>
-            )}
-          </CardContent>
-        </Card>
-
-        <Card className="border border-border/70">
-          <CardHeader>
-            <CardTitle>Observer la partie</CardTitle>
-            <CardDescription>
-              Suivez le déroulement en direct sans intervenir sur les plateaux.
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-4">
-            <p className="text-sm text-muted-foreground">
-              Visualisez les cartes retournées en temps réel et les actions des
-              joueurs sans impacter la partie. Vous pourrez tenter de rejoindre
-              plus tard si une place se libère.
-            </p>
-            {isAlreadyPlayer && !isViewingAsSpectator ? (
-              <div className="rounded-md border border-border/60 bg-muted/30 px-3 py-2 text-sm text-muted-foreground">
-                Passer en mode spectateur désactive les interactions avec votre
-                plateau sur cet appareil. Vous pourrez redevenir joueur à tout
-                moment.
-              </div>
-            ) : null}
-            <Button
-              type="button"
-              variant="outline"
-              onClick={onConfirmSpectator}
-            >
-              <EyeIcon aria-hidden className="mr-2 size-4" />
-              {isViewingAsSpectator
-                ? "Continuer en spectateur"
-                : "Observer la partie"}
-            </Button>
-          </CardContent>
-        </Card>
-      </div>
-    </div>
-  );
-}
-
-function MissingOpponentBoard({ grid }: { grid: Grid | null }) {
-  return (
-    <Card className="flex h-full flex-col border-dashed border-border/70 bg-muted/20">
-      <CardHeader>
-        <CardTitle>En attente d’un adversaire</CardTitle>
-        <CardDescription>
-          {grid
-            ? `Le plateau ${grid.rows} × ${grid.columns} est prêt à être partagé.`
-            : "Configurez un plateau pour démarrer la partie."}
-        </CardDescription>
-      </CardHeader>
-      <CardContent className="flex flex-1 flex-col justify-center gap-3">
-        <p className="text-sm text-muted-foreground">
-          Invitez un joueur à rejoindre cette salle pour afficher son plateau et
-          démarrer la partie.
-        </p>
-      </CardContent>
-    </Card>
-  );
-}
-
-function FinalResultCard({
-  state,
-  cardLookup,
-}: {
-  state: Extract<GameState, { status: typeof GameStatus.Finished }>;
-  cardLookup: Map<string, GameCard>;
-}) {
-  const winner = selectWinner(state);
-  const guess = state.finalGuess;
-  const guessedCard = cardLookup.get(guess.cardId);
-  const guesser = state.players.find((player) => player.id === guess.guesserId);
-  const target = state.players.find(
-    (player) => player.id === guess.targetPlayerId,
-  );
-
-  return (
-    <Card className="border border-emerald-400/60 bg-emerald-500/10 shadow-sm dark:border-emerald-500/40 dark:bg-emerald-500/15">
-      <CardHeader>
-        <CardTitle>
-          Victoire de {winner ? winner.name : "joueur inconnu"}
-        </CardTitle>
-        <CardDescription>{getConclusionLabel(state.reason)}</CardDescription>
-      </CardHeader>
-      <CardContent className="space-y-2 text-sm">
-        <p className="text-foreground">
-          {guesser ? guesser.name : "Un joueur"} a tenté de deviner{" "}
-          {guessedCard ? guessedCard.label : guess.cardId} chez{" "}
-          {target ? target.name : "son adversaire"}.
-        </p>
-        <p className="text-muted-foreground">
-          Tour {state.turn} —{" "}
-          {guess.correct ? "réponse correcte." : "réponse incorrecte."}
-        </p>
-      </CardContent>
-    </Card>
-  );
-}
 export default function RoomPage() {
   const params = useParams<{ roomId: string }>();
   const rawRoomId = params?.roomId;
@@ -2057,7 +706,7 @@ export default function RoomPage() {
     [players, gameState, effectiveLocalPlayerId, activePlayerId],
   );
 
-  const hostIdentityForInvite = useMemo(() => {
+  const hostIdentityForInvite = useMemo<HostIdentitySummary | null>(() => {
     if (hostPreparation) {
       return { id: hostPreparation.hostId, name: hostPreparation.nickname };
     }
